// src/app/syllabus-entry/models/TrainingObjective.ts
export interface TrainingObjective {
      trainingObjectiveId:number;
      courseTaskId: number;
      baseSchoolNameId: number;
      courseNameId: number;
      courseName:string;
      courseDurationId: number;
      bnaSubjectNameId: number;
      trainingObjectDetail: string;
      remarks:number;
      menuPosition:number;
      isActive: boolean
}